+++ packages/spotlight/CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact spotlight module, newest
 ### Fixed
 
 - `spotlight` to retry setting focus when the window is activated
+- `spotlight` handling of 5-way events after the pointer hides
 
 ## [2.0.0-beta.2] - 2018-05-07